fix: upload/import progress, serialized imports, MIME types, deployment package
- Fix upload progress bar (slotIndex for parallel file slots)
- Serialize import calls to avoid SQLite database locked error
- Fix progress bar not reaching 100% on completion
- Fix MIME type for JS files (text/plain -> application/javascript)
- Fix login_test is_delete column compatibility
- Remove pymysql from __init__.py (using SQLite not MySQL)
- Add openpyxl/xlrd to requirements.txt
- Add .gitignore for node_modules, __pycache__, dist, db.sqlite3
- Add analysis module for workload analysis
- Add deployment start.bat with auto Python detection

+++ BackEnd/scripts/import_tracks_from_csv.py
@@ -51,8 +51,8 @@ def import_all(db_path: Path, csv_dir: Path):
                 course = float(r[7]) if len(r) > 7 and r[7] else None
                 workstatus = 1 if (len(r) > 8 and (r[8].strip().upper() in ('TRUE','1','T')) ) else 0
                 width = float(r[9]) if len(r) > 9 and r[9] else None
-                depth = int(r[10]) if len(r) > 10 and r[10] else None
-                depthstandard = int(r[11]) if len(r) > 11 and r[11] else None
+                depth = float(r[10]) if len(r) > 10 and r[10] else None
+                depthstandard = float(r[11]) if len(r) > 11 and r[11] else None
             except Exception:
                 gpstime = None
                 lon = lat = x = y = velocity = course = None
